pyproject.toml
[project]
name = "ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = "==3.11.*"
dependencies = [
    "aiokafka>=0.12.0",
    "asyncmy>=0.2.11",
    "beautifulsoup4>=4.12.3",
    "boto3>=1.42.37",
    "fastapi>=0.115.6",
    "kafka-python>=2.0.2",
    "langchain>=0.3.14",
    "langchain-community>=0.3.14",
    "langchain-core>=0.3.29",
    "langchain-google-genai>=4.2.0",
    "langchain-openai>=0.2.14",
    "langchain-text-splitters>=0.3.5",
    "langgraph>=0.2.61",
    "pdfplumber>=0.11.9",
    "playwright>=1.57.0",
    "psycopg2-binary>=2.9.10",
    "pydantic-settings>=2.7.1",
    "pypdf>=6.6.2",
    "python-dotenv>=1.0.1",
    "pyvirtualdisplay>=3.0",
    "sqlalchemy>=2.0.37",
    "taskiq>=0.11.10",
    "taskiq-aio-pika>=0.4.1",
    "uvicorn>=0.34.0",
    "weaviate-client>=4.9.3",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = [
    "api", 
    "shared",
    "pipelines/job_analysis",
    "pipelines/resume_analysis",
    "pipelines/applicant_evaluation",
    "pipelines/candidate_comparison",
]

[dependency-groups]
dev = [
    "black>=26.1.0",
    "mypy>=1.19.1",
    "pytest>=8.3.4",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.9.2",
]

[tool.mypy]
explicit_package_bases = true
namespace_packages = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = [
    "tests/unit"
]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
